+++ src/components/EmojiPicker/EmojiPickerMenu/BaseEmojiPickerMenu.js
@@ -5,16 +5,19 @@ import {Text, View} from 'react-native';
 import CategoryShortcutBar from '@components/EmojiPicker/CategoryShortcutBar';
 import EmojiSkinToneList from '@components/EmojiPicker/EmojiSkinToneList';
 import refPropTypes from '@components/refPropTypes';
-import withLocalize, {withLocalizePropTypes} from '@components/withLocalize';
+import useLocalize from '@hooks/useLocalize';
 import stylePropTypes from '@styles/stylePropTypes';
 import useThemeStyles from '@styles/useThemeStyles';
 import CONST from '@src/CONST';
-import getItemType from './getItemType';
 
-const emojiPropType = {
+const emojiPropTypes = {
+    /** The code of the item */
     code: PropTypes.string.isRequired,
+    /** Whether the item is a header or not */
     header: PropTypes.bool,
+    /** Whether the item is a spacer or not */
     spacer: PropTypes.bool,
+    /** Types of an emoji - e.g. different skin types */
     types: PropTypes.arrayOf(PropTypes.string),
 };
 
@@ -23,7 +26,7 @@ const propTypes = {
     isFiltered: PropTypes.bool.isRequired,
 
     /** Array of header emojis */
-    headerEmojis: PropTypes.arrayOf(PropTypes.shape(emojiPropType)).isRequired,
+    headerEmojis: PropTypes.arrayOf(PropTypes.shape(emojiPropTypes)).isRequired,
 
     /** Function to scroll to a specific header in the emoji list */
     scrollToHeader: PropTypes.func.isRequired,
@@ -35,13 +38,10 @@ const propTypes = {
     forwardedRef: refPropTypes,
 
     /** The data for the emoji list */
-    data: PropTypes.arrayOf(PropTypes.shape(emojiPropType)).isRequired,
+    data: PropTypes.arrayOf(PropTypes.shape(emojiPropTypes)).isRequired,
 
     /** Function to render each item in the list */
     renderItem: PropTypes.func.isRequired,
-
-    /** Function to extract a unique key for each item in the list */
-    keyExtractor: PropTypes.func.isRequired,
 
     /** Extra data to be passed to the list for re-rendering */
     // eslint-disable-next-line react/forbid-prop-types
@@ -58,8 +58,6 @@ const propTypes = {
 
     /** Whether the list should always bounce vertically */
     alwaysBounceVertical: PropTypes.bool,
-
-    ...withLocalizePropTypes,
 };
 
 const defaultProps = {
@@ -71,6 +69,36 @@ const defaultProps = {
     alwaysBounceVertical: false,
 };
 
+/**
+ * Improves FlashList's recycling when there are different types of items
+ * @param {Object} item
+ * @returns {String}
+ */
+const getItemType = (item) => {
+    // item is undefined only when list is empty
+    if (!item) {
+        return;
+    }
+
+    if (item.name) {
+        return CONST.EMOJI_PICKER_ITEM_TYPES.EMOJI;
+    }
+    if (item.header) {
+        return CONST.EMOJI_PICKER_ITEM_TYPES.HEADER;
+    }
+
+    return CONST.EMOJI_PICKER_ITEM_TYPES.SPACER;
+};
+
+/**
+ * Return a unique key for each emoji item
+ *
+ * @param {Object} item
+ * @param {Number} index
+ * @returns {String}
+ */
+const keyExtractor = (item, index) => `emoji_picker_${item.code}_${index}`;
+
 function BaseEmojiPickerMenu({
     headerEmojis,
     scrollToHeader,
@@ -79,15 +107,14 @@ function BaseEmojiPickerMenu({
     forwardedRef,
     data,
     renderItem,
-    keyExtractor,
     stickyHeaderIndices,
     onUpdatePreferredSkinTone,
     preferredSkinTone,
     extraData,
     alwaysBounceVertical,
-    translate,
 }) {
     const styles = useThemeStyles();
+    const {translate} = useLocalize();
 
     return (
         <>
@@ -136,4 +163,4 @@ const BaseEmojiPickerMenuWithRef = React.forwardRef((props, ref) => (
 
 BaseEmojiPickerMenuWithRef.displayName = 'BaseEmojiPickerMenuWithRef';
 
-export default withLocalize(BaseEmojiPickerMenuWithRef);
+export default BaseEmojiPickerMenuWithRef;
